Refactorizado de nombres de variables y metodos

+++ src/main/java/es/hefame/keyvault/datastructure/message/InfoMessage.java
@@ -23,7 +23,7 @@ public class InfoMessage implements JsonEncodable {
 	public JSONAware jsonEncode() {
 		JSONObject root = new JSONObject();
 		if (code != null) {
-			root.put("code", code.intValue());
+			root.put("code", code);
 		}
 		root.put("message", message);
 		return root;
